# trie/trienode: minor speedup in nodeset merging