Shaves off a bit of alloc during nodes merge
```
name         old time/op    new time/op    delta
Merge/1K-8      600µs ± 7%     544µs ± 4%   -9.26%  (p=0.000 n=9+10)
Merge/10K-8    5.99ms ± 8%    5.48ms ± 7%   -8.64%  (p=0.000 n=10+10)

name         old alloc/op   new alloc/op   delta
Merge/1K-8      252kB ± 0%     244kB ± 0%   -3.17%  (p=0.000 n=8+10)
Merge/10K-8    2.02MB ± 0%    1.94MB ± 0%   -3.96%  (p=0.000 n=10+10)

name         old allocs/op  new allocs/op  delta
Merge/1K-8      2.08k ± 0%     0.08k ± 0%  -96.29%  (p=0.000 n=10+9)
Merge/10K-8     20.6k ± 0%      0.6k ± 1%  -97.31%  (p=0.000 n=10+10)
```